refactor: dto에 toEntity()를 만들어 변환 (#8)
