--- src/main/java/com/example/talkenbackend/feedback/dto/FeedbackRequestDto.java
@@ -1,13 +1,0 @@
-package com.example.talkenbackend.feedback.dto;
-
-import lombok.Data;
-
-@Data
-public class FeedbackRequestDto {
-    private Long mentorId;
-    private Long menteeId;
-    private Long portfolioId;
-
-    private String content;
-
-}
