+++ src/main/java/com/example/talkenbackend/feedback/service/FeedbackService.java
@@ -1,8 +1,8 @@
 package com.example.talkenbackend.feedback.service;
 
 import com.example.talkenbackend.feedback.domain.Feedback;
-import com.example.talkenbackend.feedback.dto.FeedbackRequestDto;
-import com.example.talkenbackend.feedback.dto.FeedbackResponseDto;
+import com.example.talkenbackend.feedback.dto.request.FeedbackRequestDto;
+import com.example.talkenbackend.feedback.dto.response.FeedbackResponseDto;
 import com.example.talkenbackend.feedback.exception.FeedbackNotFoundException;
 import com.example.talkenbackend.feedback.repository.FeedbackRepository;
 import lombok.RequiredArgsConstructor;
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static com.example.talkenbackend.feedback.dto.FeedbackResponseDto.fromEntity;
+import static com.example.talkenbackend.feedback.dto.response.FeedbackResponseDto.fromEntity;
 
 @RequiredArgsConstructor
 
@@ -29,13 +29,11 @@ public class FeedbackService {
     public List<FeedbackResponseDto> getFeedbackByMenteeId(Long menteeId) throws Exception {
         return feedbackRepository.findByMenteeId(menteeId).orElseThrow(() -> new Exception())
                 .stream().map(x -> fromEntity(x)).collect(Collectors.toList());
-
     }
 
     public List<FeedbackResponseDto> getFeedbackByMentorId(Long mentorId) throws Exception {
         return feedbackRepository.findByMentorId(mentorId).orElseThrow(() -> new Exception())
                 .stream().map(x -> fromEntity(x)).collect(Collectors.toList());
-
     }
 
     @Transactional
@@ -44,25 +42,14 @@ public class FeedbackService {
     }
 
     public FeedbackResponseDto createFeedback(FeedbackRequestDto dto) {
-        Feedback feedback = Feedback.builder()
-                .portfolioId(dto.getPortfolioId())
-                .menteeId(dto.getMenteeId())
-                .mentorId(dto.getMentorId())
-                .content(dto.getContent())
-                .build();
-
-        return fromEntity(feedbackRepository.save(feedback));
+        return fromEntity(feedbackRepository.save(dto.toEntity()));
     }
 
     @Transactional
     public FeedbackResponseDto updateFeedback(Long feedbackId, FeedbackRequestDto dto) {
         Feedback feedback = feedbackRepository.findById(feedbackId).orElseThrow(
-                () -> new FeedbackNotFoundException(feedbackId)
-        );
-
+                () -> new FeedbackNotFoundException(feedbackId));
         feedback.update(dto);
-//        Feedback save = feedbackRepository.save(feedback);
-
         return fromEntity(feedback);
     }
 }
